pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "streamlit-extras>=0.6.0",
    "streamlit>=1.44.1",
    "streamlit-lottie>=0.0.5",
    "requests>=2.32.3",
    "anthropic>=0.49.0",
]
